Send tagging notifications asynchronously
