+++ lib/indexer/metadata_tagger.rb
@@ -1,5 +1,5 @@
 require 'csv'
-require 'gds_api/email_alert_api'
+require 'indexer/workers/metadata_tagger_notification_worker'
 
 module Indexer
   class MetadataTagger
@@ -32,7 +32,7 @@ module Indexer
           Indexer::AmendWorker.new.perform(index_to_update, base_path, metadata)
 
           unless base_paths.empty? || base_paths.include?(base_path)
-            send_notification(item_in_search["_source"], metadata)
+            Indexer::MetadataTaggerNotificationWorker.perform_async(item_in_search, metadata)
           end
         end
       end
@@ -93,48 +93,6 @@ module Indexer
 
     def self.all_indexed_eu_exit_guidance_paths
       find_all_eu_exit_guidance[:results].collect { |r| r["link"] }
-    end
-
-    def self.send_notification(document, metadata)
-      payload = email_alert_api_payload(document, metadata)
-
-      begin
-        email_alert_api.send_alert(payload)
-        puts "notification sent for #{payload}"
-      rescue GdsApi::HTTPConflict
-        puts "email-alert-api returned 409 conflict for #{payload}"
-      end
-    end
-
-    def self.email_alert_api_payload(document, metadata)
-      {
-        title: document["title"],
-        description: document["description"],
-        change_note: "This publication has just been added to the EU Exit business guidance finder on GOV.UK.",
-        subject: document["title"],
-        tags: metadata,
-        links: {
-          content_id: document["content_id"],
-          organisations: document["organisation_content_ids"],
-          taxons: document["taxons"],
-        },
-        urgent: true,
-        document_type: document["content_store_document_type"],
-        email_document_supertype: "other",
-        government_document_supertype: "other",
-        content_id: document["content_id"],
-        public_updated_at: document["public_timestamp"],
-        publishing_app: document.fetch("publishing_app", "rummager"),
-        base_path: document["link"],
-        priority: "high",
-      }
-    end
-
-    def self.email_alert_api
-      @email_alert_api ||= GdsApi::EmailAlertApi.new(
-        Plek.current.find('email-alert-api'),
-        bearer_token: ENV['EMAIL_ALERT_API_BEARER_TOKEN'] || 'example123'
-      )
     end
   end
 end
